+++ conference-xml/src/main/java/learn/springpath/conference/Application.java
@@ -1,12 +1,14 @@
 package learn.springpath.conference;
 
 import learn.springpath.conference.service.SpeakerService;
-import learn.springpath.conference.service.SpeakerServiceImpl;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 public class Application {
 
     public static void main(String[] args) {
-        SpeakerService service = new SpeakerServiceImpl();
+        ApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+        SpeakerService service = context.getBean("speakerService", SpeakerService.class);
         System.out.println(service.findAll());
     }
 }