+++ conference-xml/src/main/java/learn/springpath/conference/service/SpeakerServiceImpl.java
@@ -1,14 +1,17 @@
 package learn.springpath.conference.service;
 
 import learn.springpath.conference.model.Speaker;
-import learn.springpath.conference.repository.HibernateSpeakerRepositoryImpl;
 import learn.springpath.conference.repository.SpeakerRepository;
 
 import java.util.List;
 
 public class SpeakerServiceImpl implements SpeakerService {
 
-    private final SpeakerRepository speakerRepository = new HibernateSpeakerRepositoryImpl();
+    private SpeakerRepository speakerRepository;
+
+    public void setSpeakerRepository(SpeakerRepository speakerRepository) {
+        this.speakerRepository = speakerRepository;
+    }
 
     @Override
     public List<Speaker> findAll() {