+++ conference-xml/History.md
@@ -1,5 +1,9 @@
 # History
 
+### 2 - Add XML configuration
+* Config beans in applicationContext.xml
+* Use ClassPathXmlApplicationContext to bootstrap application
+
 ### 1 - Add Model, Repository and Service for Speaker (without Spring)
 * Speaker
 * SpeakerRepository